feat: separating layouts to login and dashboard
