+++ src/app/dashboard/page.tsx
@@ -1,3 +1,3 @@
 export default function DashboadPage() {
-  return <h1>dashboard</h1>;
+  return <main className="main-dashboard">DASHBOARD PAGE</main>;
 }
