+++ src/app/globals.scss
@@ -2,41 +2,60 @@
 @tailwind components;
 @tailwind utilities;
 @import "variables";
-@import url('https://fonts.googleapis.com/css2?family=Nunito+Sans:ital,opsz,wght@0,6..12,200..1000;1,6..12,200..1000&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Nunito+Sans:ital,opsz,wght@0,6..12,200..1000;1,6..12,200..1000&display=swap");
 
 body {
   font-family: "Nunito Sans", sans-serif;
   padding: 0;
   margin: 0;
   box-sizing: border-box;
-  background-color: #ff7672;
-  background-image: linear-gradient(45deg, #ff7672 5%, #ffffff 89%);
 }
 
-.login-box {
-  height: 65vh;
-  width: 30vw;
-  border-radius: 15px;
-  background-color: #ffffff;
-  border: 4px solid var(--red-border);
-  .header {
-    height: 30%;
+//Login Page Styles
+.bg-login {
+  background-image: url("/assets/fresh-salad-plate.jpg");
+  background-size: cover;
+  .mask {
     width: 100%;
-    .logo {
-      height: 8rem;
-      width: 8rem;
+    height: 100%;
+    background-color: rgba($color: #494a4e, $alpha: 0.65);
+    .login-box {
+      height: 65vh;
+      width: 30vw;
+      border-radius: 15px;
+      background-color: #ffffff;
+      border: 4px solid var(--red-border);
+      .header {
+        height: 30%;
+        width: 100%;
+        .logo {
+          height: 8rem;
+          width: 8rem;
+        }
+      }
+      .body {
+        height: 65%;
+        width: 100%;
+        h1 {
+          font-size: 1.25rem;
+          color: var(--red-primary);
+          font-weight: 600;
+        }
+      }
     }
-  }
-  .body {
-    height: 65%;
-    width: 100%;
-    h1{
-      font-size: 1.25rem;
-      color: var(--red-primary);
-      font-weight: 600;
+    .toast-user {
+      position: absolute;
+      bottom: 4%;
+      right: 2%;
+      background-color: #ffffff;
+      border: 2px solid var(--red-border);
+      padding: 1rem 1rem 1rem 1rem;
+      border-radius: 7px;
     }
   }
 }
+
+// Default Styles
 .label {
   font-size: 1.05rem;
   color: var(--color-text);
